test(logic): split combat_resolver part2 spy/civilian tests (Refs #2288)
Split the resolveBattleContext Spy timer interaction group (plus the
collocated general medals morale test) out of combat_resolver_test_part2_test.dart
into a new combat_resolver_test_part2_spy_civilian_test.dart so each file
stays under the repo.logic_test_file_size 400-line gate.

Both new files contain only structural test reorganization; no test cases,
assertions, or fixtures were changed.
